+++ tree.jquery.debug.js
@@ -1,5 +1,5 @@
 /*
-JqTree 1.9.1
+JqTree 2.0.1
 
 Copyright 2026 Marco Braak
 
@@ -4238,7 +4238,7 @@ var jqtree = (function (exports) {
     return !event.isDefaultPrevented();
   };
 
-  const version = "1.9.1";
+  const version = "2.0.1";
 
   const NODE_PARAM_IS_EMPTY = "Node parameter is empty";
   const PARAM_IS_EMPTY = "Parameter is empty: ";
@@ -4516,9 +4516,13 @@ var jqtree = (function (exports) {
       const jqTreeOnLoadFailed = this._inputOptions.onLoadFailed;
       if (jqTreeOnLoadFailed) {
         const {
+          error,
           response
         } = e;
-        jqTreeOnLoadFailed(response);
+        jqTreeOnLoadFailed({
+          error,
+          response
+        });
       }
     }
     _handleLoadingDataEvent(e) {